create FeedStoreSpecs that must be implemented by share scope so it can be used by other `FeedStore` implementation tests
