+++ PokemonEncyclopedia/PokemonEncyclopediaTests/FeedCache/CodableFeedStoreTests.swift
@@ -8,41 +8,7 @@
 import XCTest
 import PokemonEncyclopedia
 
-protocol FeedStoreSpecs {
-	func test_retrieve_deliversEmptyOnEmptyCache()
-	func test_retrieve_hasNoSideEffectsOnEmptyCache()
-	func test_retrieve_deliversFoundValuesOnNonEmptyCache()
-	func test_retrieve_hasNoSideEffectsOnNonEmptyCache()
-	
-	func test_insert_deliversNoErrorOnEmptyCache()
-	func test_insert_deliversNoErrorOnNonEmptyCache()
-	func test_insert_overridesPreviouslyInsertedCacheValues()
-	
-	func test_delete_deliversNoErrorOnEmptyCache()
-	func test_delete_hasNoSideEffectsOnEmptyCache()
-	func test_delete_deliversNoErrorOnNonEmptyCache()
-	func test_delete_emptiesPreviouslyInsertedCache()
-	
-	func test_storeSideEffects_runSerially()
-}
-
-protocol FailableRetrieveFeedStoreSpecs {
-	func test_retrieve_deliversFailureOnRetrievalError()
-	func test_retrieve_hasNoSideEffectsOnFailure()
-}
-
-protocol FailableInsertFeedStoreSpecs {
-	func test_insert_deliversErrorOnInsertionError()
-	func test_insert_hasNoSideEffectOnInsertionError()
-}
-
-protocol FailableDeleteFeedStoreSpecs {
-	func test_delete_deliversErrorOnDeletionError()
-	func test_delete_hasNoSideEffectOnDeletionError()
-}
-
-
-class CodableFeedStoreTests: XCTestCase {
+class CodableFeedStoreTests: XCTestCase, FailableFeedStore{
 
 	override func setUp() {
 		super.setUp()
