+++ PokemonEncyclopedia/PokemonEncyclopediaTests/FeedCache/FeedStoreSpecs.swift
@@ -1,0 +1,44 @@
+//
+//  FeedStoreSpecs.swift
+//  PokemonEncyclopediaTests
+//
+//  Created by Ankit on 19/07/21.
+//
+
+import Foundation
+
+protocol FeedStoreSpecs {
+	func test_retrieve_deliversEmptyOnEmptyCache()
+	func test_retrieve_hasNoSideEffectsOnEmptyCache()
+	func test_retrieve_deliversFoundValuesOnNonEmptyCache()
+	func test_retrieve_hasNoSideEffectsOnNonEmptyCache()
+	
+	func test_insert_deliversNoErrorOnEmptyCache()
+	func test_insert_deliversNoErrorOnNonEmptyCache()
+	func test_insert_overridesPreviouslyInsertedCacheValues()
+	
+	func test_delete_deliversNoErrorOnEmptyCache()
+	func test_delete_hasNoSideEffectsOnEmptyCache()
+	func test_delete_deliversNoErrorOnNonEmptyCache()
+	func test_delete_emptiesPreviouslyInsertedCache()
+	
+	func test_storeSideEffects_runSerially()
+}
+
+protocol FailableRetrieveFeedStoreSpecs: FeedStoreSpecs {
+	func test_retrieve_deliversFailureOnRetrievalError()
+	func test_retrieve_hasNoSideEffectsOnFailure()
+}
+
+protocol FailableInsertFeedStoreSpecs: FeedStoreSpecs {
+	func test_insert_deliversErrorOnInsertionError()
+	func test_insert_hasNoSideEffectOnInsertionError()
+}
+
+protocol FailableDeleteFeedStoreSpecs: FeedStoreSpecs {
+	func test_delete_deliversErrorOnDeletionError()
+	func test_delete_hasNoSideEffectOnDeletionError()
+}
+
+
+typealias FailableFeedStore = FailableRetrieveFeedStoreSpecs & FailableInsertFeedStoreSpecs & FailableDeleteFeedStoreSpecs
